+++ include/scheduler/scheduler.h
@@ -2,16 +2,19 @@
 #define CAVALIERI_SCHEDULER_SCHEDULER_H
 
 #include <functional>
-#include <memory>
+#include <future>
 #include <time.h>
 #include <async/async_loop.h>
 
-typedef std::function<void()> task_fn_t;
+using task_fn_t = std::function<void()>;
+using remove_task_fn_t = std::function<void()>;
+using remove_task_future_t = std::future<remove_task_fn_t>;
 
 class scheduler_interface {
 public:
-  virtual void add_periodic_task(task_fn_t task, float interval) = 0;
-  virtual void add_once_task(task_fn_t task, float dt) = 0;
+  virtual remove_task_future_t add_periodic_task(task_fn_t task,
+                                                  float interval) = 0;
+  virtual remove_task_future_t add_once_task(task_fn_t task, float dt) = 0;
   virtual time_t unix_time() = 0;
   virtual void set_time(const time_t t) = 0;
   virtual void clear() = 0;